withdraw amount from customer and add 70 percent for specialist;

+++ src/main/java/org/example/serviceapplication/credit/model/CreditDto.java
@@ -7,8 +7,8 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public record CreditDto(
-         Integer userId,
-         Integer balance
+         Long userId,
+         Double balance
 ) {
 
 }
